Rename volara.tmp to volara.segment_utils, keep a deprecating alias
`volara/tmp.py` was named like scratch code but is load-bearing public API:
`seg_to_affgraph` appears in the published tutorial, and the numba kernels
(`replace_values`, `replace_values_sorted`, `prepare_mapping`,
`filter_mapping_to_block`, `warmup_replace_values_sorted`) are imported by
`volara/lut.py` and three `volara/blockwise` tasks. The name told readers it was
safe to delete, which it is not.

Renamed to `volara/segment_utils.py` -- every function in it operates on
segmentation label arrays -- and updated all in-repo importers plus
`docs/source/tutorial.rst`.

`volara/tmp.py` remains as a thin alias that re-exports the same six objects and
emits a `DeprecationWarning`, because the old path is baked into the published
docs and may be used by downstream code. The re-exports are the identical
objects, so the numba dispatchers are shared rather than recompiled.

The module body is a byte-identical move; no behaviour changed.

`volara/__init__.py`'s lazy `__getattr__` only lists `datasets` and `dbs`, so it
needed no update.
