+++ tests/conftest.py
@@ -8,7 +8,7 @@ from funlib.persistence.arrays import prepare_ds
 
 from volara.dbs import SQLite
 from volara.logging import set_log_basedir
-from volara.tmp import seg_to_affgraph
+from volara.segment_utils import seg_to_affgraph
 
 
 @pytest.fixture(autouse=True)
